refactor: split opencode plugin index into focused modules
