+++ .agents/plugins/opencode-aidevops/index.mjs
@@ -1,7 +1,6 @@
 import { execSync } from "child_process";
 import {
   readFileSync,
-  readdirSync,
   existsSync,
   appendFileSync,
   mkdirSync,
@@ -13,6 +12,10 @@ import { join } from "path";
 import { homedir, platform } from "os";
 import { createTools } from "./tools.mjs";
 import { initObservability, handleEvent, recordToolCall } from "./observability.mjs";
+import { loadAgentIndex, applyAgentMcpTools } from "./agent-loader.mjs";
+import { validateReturnStatements, validatePositionalParams } from "./validators.mjs";
+import { runMarkdownQualityPipeline } from "./quality-pipeline.mjs";
+import { createTtsrHooks } from "./ttsr.mjs";
 
 const HOME = homedir();
 const AGENTS_DIR = join(HOME, ".aidevops", "agents");
@@ -63,41 +66,6 @@ function readIfExists(filepath) {
   return "";
 }
 
-/**
- * Parse YAML frontmatter from a markdown file.
- * Lightweight parser — no dependencies. Handles the common cases.
- * @param {string} content
- * @returns {{ data: Record<string, any>, body: string }}
- */
-function parseFrontmatter(content) {
-  const match = content.match(/^---\r?\n([\s\S]*?)\r?\n---\r?\n?([\s\S]*)$/);
-  if (!match) return { data: {}, body: content };
-
-  const yaml = match[1];
-  const body = match[2];
-  const data = {};
-
-  for (const line of yaml.split("\n")) {
-    const trimmed = line.trim();
-    if (!trimmed || trimmed.startsWith("#")) continue;
-
-    const colonIdx = trimmed.indexOf(":");
-    if (colonIdx === -1) continue;
-
-    const key = trimmed.slice(0, colonIdx).trim();
-    let value = trimmed.slice(colonIdx + 1).trim();
-
-    // Handle booleans and numbers
-    if (value === "true") value = true;
-    else if (value === "false") value = false;
-    else if (/^\d+(\.\d+)?$/.test(value)) value = Number(value);
-
-    data[key] = value;
-  }
-
-  return { data, body };
-}
-
 // ---------------------------------------------------------------------------
 // Phase 1: Lightweight Agent Discovery (t1040)
 // ---------------------------------------------------------------------------
@@ -111,152 +79,7 @@ function parseFrontmatter(content) {
 // The index is manually maintained in the repo. The fallback ensures new agents
 // are still discoverable even if the index is stale or missing.
 
-/** Names to skip when discovering agents. */
-const SKIP_NAMES = new Set([
-  "README",
-  "AGENTS",
-  "SKILL",
-  "SKILL-SCAN-RESULTS",
-  "node_modules",
-  "references",
-  "loop-state",
-]);
-
-/**
- * Collect leaf agent names from a pipe-separated key_files string.
- * @param {string} keyFiles - e.g. "dataforseo|serper|semrush"
- * @param {string} purpose - Description for the agent entry
- * @param {Array} agents - Mutable agents array
- * @param {Set} seen - Dedup set
- */
-function collectLeafAgents(keyFiles, purpose, agents, seen) {
-  for (const leaf of keyFiles.split("|")) {
-    const name = leaf.trim();
-    if (!name || SKIP_NAMES.has(name) || name.endsWith("-skill")) continue;
-    if (seen.has(name)) continue;
-    seen.add(name);
-    agents.push({ name, description: purpose });
-  }
-}
-
-/**
- * Parse a TOON subagents block into agent entries.
- * Each line: folder,purpose,keyfile1|keyfile2|...
- * @param {string} blockText - Raw text from the TOON block
- * @returns {Array<{name: string, description: string}>}
- */
-function parseToonSubagentBlock(blockText) {
-  const agents = [];
-  const seen = new Set();
-
-  for (const line of blockText.split("\n")) {
-    const trimmed = line.trim();
-    if (!trimmed) continue;
-
-    const parts = trimmed.split(",");
-    if (parts.length < 3) continue;
-
-    const folder = parts[0] || "";
-    if (folder.includes("references/") || folder.includes("loop-state/")) continue;
-
-    collectLeafAgents(parts.slice(2).join(","), parts[1] || "", agents, seen);
-  }
-
-  return agents;
-}
-
-/**
- * Parse subagent-index.toon and return leaf agent names with descriptions.
- * Reads ONE file instead of 500+. Returns entries like:
- *   { name: "dataforseo", description: "Search optimization - keywords..." }
- * @returns {Array<{name: string, description: string}>}
- */
-function loadAgentIndex() {
-  const indexPath = join(AGENTS_DIR, "subagent-index.toon");
-  const content = readIfExists(indexPath);
-  if (!content) return loadAgentsFallback();
-
-  const subagentMatch = content.match(
-    /<!--TOON:subagents\[\d+\]\{[^}]+\}:\n([\s\S]*?)-->/,
-  );
-  if (!subagentMatch) return loadAgentsFallback();
-
-  return parseToonSubagentBlock(subagentMatch[1]);
-}
-
-/**
- * Fallback: discover agents from directory names only (no file reads).
- * Lists .md filenames in known subdirectories — O(n) readdirSync calls
- * where n = number of subdirectories (~11), NOT number of files.
- * Each readdirSync returns filenames without reading file contents.
- * @returns {Array<{name: string, description: string}>}
- */
-function loadAgentsFallback() {
-  if (!existsSync(AGENTS_DIR)) return [];
-
-  const subdirs = [
-    "aidevops",
-    "content",
-    "seo",
-    "tools",
-    "services",
-    "workflows",
-    "memory",
-    "custom",
-    "draft",
-  ];
-
-  const agents = [];
-  const seen = new Set();
-
-  for (const subdir of subdirs) {
-    scanDirNames(join(AGENTS_DIR, subdir), subdir, agents, seen);
-  }
-
-  return agents;
-}
-
-/**
- * Recursively collect .md filenames from a directory tree.
- * Only calls readdirSync (directory listing) — never reads file contents.
- * @param {string} dirPath
- * @param {string} folderDesc - used as description fallback
- * @param {Array} agents
- * @param {Set} seen - dedup set
- */
-/**
- * Try to register a .md file entry as a discovered agent.
- * @param {object} entry - Dirent object
- * @param {string} folderDesc - Description fallback
- * @param {Array} agents - Mutable agents array
- * @param {Set} seen - Dedup set
- */
-function tryRegisterMdAgent(entry, folderDesc, agents, seen) {
-  if (!entry.isFile() || !entry.name.endsWith(".md")) return;
-  const name = entry.name.replace(/\.md$/, "");
-  if (SKIP_NAMES.has(name) || name.endsWith("-skill")) return;
-  if (seen.has(name)) return;
-  seen.add(name);
-  agents.push({ name, description: `aidevops subagent: ${folderDesc}` });
-}
-
-function scanDirNames(dirPath, folderDesc, agents, seen) {
-  let entries;
-  try {
-    entries = readdirSync(dirPath, { withFileTypes: true });
-  } catch {
-    return;
-  }
-
-  for (const entry of entries) {
-    if (entry.isDirectory()) {
-      if (SKIP_NAMES.has(entry.name)) continue;
-      scanDirNames(join(dirPath, entry.name), folderDesc, agents, seen);
-    } else {
-      tryRegisterMdAgent(entry, folderDesc, agents, seen);
-    }
-  }
-}
+// Agent index loading extracted to agent-loader.mjs
 
 // ---------------------------------------------------------------------------
 // Phase 2: MCP Server Registry + Config Hook
@@ -417,32 +240,6 @@ function getMcpRegistry() {
 }
 
 /**
- * Map of subagent names to the MCP tool patterns they need enabled.
- * Used by the config hook to set per-agent tool permissions.
- *
- * Only includes subagents that need MCP tools beyond the defaults.
- * Agents not listed here get only the globally-enabled tools.
- */
-const AGENT_MCP_TOOLS = {
-  outscraper: ["outscraper_*"],
-  mainwp: ["localwp_*"],
-  localwp: ["localwp_*"],
-  quickfile: ["quickfile_*"],
-  "google-search-console": ["gsc_*"],
-  dataforseo: ["dataforseo_*"],
-  "claude-code": ["claude-code-mcp_*"],
-  playwriter: ["playwriter_*"],
-  shadcn: ["shadcn_*"],
-  "macos-automator": IS_MACOS ? ["macos-automator_*"] : [],
-  mac: IS_MACOS ? ["macos-automator_*"] : [],
-  "ios-simulator-mcp": IS_MACOS ? ["ios-simulator_*"] : [],
-  "augment-context-engine": ["augment-context-engine_*"],
-  context7: ["context7_*"],
-  sentry: ["sentry_*"],
-  socket: ["socket_*"],
-};
-
-/**
  * Check if an MCP entry should be skipped (wrong platform, missing binary).
  * @param {object} mcp - MCP registry entry
  * @param {object} tools - Config tools object (mutable — disables pattern if binary missing)
@@ -530,57 +327,7 @@ function registerMcpServers(config) {
   return registered;
 }
 
-/**
- * Apply per-agent MCP tool permissions.
- * Ensures subagents that need specific MCP tools have them enabled
- * in their agent config, even if the tools are disabled globally.
- *
- * @param {object} config - OpenCode Config object (mutable)
- * @returns {number} Number of agents updated
- */
-/**
- * Apply tool patterns to a single agent config entry.
- * Only sets tools not already configured (shell script takes precedence).
- * @param {object} agentEntry - Mutable agent config object
- * @param {string[]} toolPatterns - Tool patterns to enable
- * @returns {number} Number of tools newly enabled
- */
-function applyToolPatternsToAgent(agentEntry, toolPatterns) {
-  let count = 0;
-  if (!agentEntry.tools) {
-    agentEntry.tools = {};
-  }
-  for (const pattern of toolPatterns) {
-    if (!(pattern in agentEntry.tools)) {
-      agentEntry.tools[pattern] = true;
-      count++;
-    }
-  }
-  return count;
-}
-
-function applyAgentMcpTools(config) {
-  if (!config.agent) return 0;
-
-  let updated = 0;
-
-  for (const [mcpAgentName, toolPatterns] of Object.entries(AGENT_MCP_TOOLS)) {
-    if (toolPatterns.length === 0) continue;
-
-    // Find matching agent(s) — check both exact name and path-based names
-    // ending with the basename (t1015: agents now use path-based names like
-    // "tools/wordpress/mainwp" instead of just "mainwp")
-    const matchingKeys = Object.keys(config.agent).filter(
-      (key) => key === mcpAgentName || key.endsWith("/" + mcpAgentName),
-    );
-
-    for (const matchKey of matchingKeys) {
-      updated += applyToolPatternsToAgent(config.agent[matchKey], toolPatterns);
-    }
-  }
-
-  return updated;
-}
+// Agent MCP tool permissions extracted to agent-loader.mjs
 
 /**
  * Modify OpenCode config to register aidevops subagents, MCP servers,
@@ -596,7 +343,7 @@ async function configHook(config) {
   if (!config.agent) config.agent = {};
 
   // --- Lightweight agent registration from pre-built index ---
-  const indexAgents = loadAgentIndex();
+  const indexAgents = loadAgentIndex(AGENTS_DIR, readIfExists);
   let agentsInjected = 0;
 
   for (const agent of indexAgents) {
@@ -684,241 +431,7 @@ function qualityDetailLog(label, filePath, report) {
   }
 }
 
-/**
- * Try to match a shell function definition on a line.
- * @param {string} trimmed - Trimmed line content
- * @returns {string|null} Function name if matched, null otherwise
- */
-function matchFunctionDef(trimmed) {
-  const funcMatch = trimmed.match(/^([a-zA-Z_][a-zA-Z0-9_]*)\s*\(\)\s*\{/);
-  if (funcMatch) return funcMatch[1];
-
-  const funcMatch2 = trimmed.match(/^function\s+([a-zA-Z_][a-zA-Z0-9_]*)/);
-  if (funcMatch2) return funcMatch2[1];
-
-  return null;
-}
-
-/**
- * Count brace depth change in a line.
- * @param {string} trimmed - Trimmed line content
- * @returns {number} Net brace depth change
- */
-function braceDepthDelta(trimmed) {
-  let delta = 0;
-  for (const ch of trimmed) {
-    if (ch === "{") delta++;
-    else if (ch === "}") delta--;
-  }
-  return delta;
-}
-
-/**
- * Check if a line contains a shell return statement.
- * @param {string} trimmed - Trimmed line content
- * @returns {boolean}
- */
-function hasReturnStatement(trimmed) {
-  return /\breturn\s+[0-9]/.test(trimmed) || /\breturn\s*$/.test(trimmed);
-}
-
-/**
- * Record a missing-return violation.
- * @param {string[]} details - Mutable details array
- * @param {number} functionStart - 1-based line number
- * @param {string} functionName
- * @returns {number} 1 (violation count increment)
- */
-function recordMissingReturn(details, functionStart, functionName) {
-  details.push(
-    `  Line ${functionStart}: function '${functionName}' missing explicit return`,
-  );
-  return 1;
-}
-
-/**
- * Check if the current function (being tracked) is missing a return and record it.
- * @param {object} state - Mutable function-tracking state
- * @param {string[]} details - Mutable details array
- * @returns {number} 0 or 1
- */
-function checkAndRecordMissingReturn(state, details) {
-  if (state.inFunction && !state.hasReturn) {
-    return recordMissingReturn(details, state.functionStart, state.functionName);
-  }
-  return 0;
-}
-
-/**
- * Begin tracking a new function definition.
- * @param {object} state - Mutable function-tracking state
- * @param {string} name - Function name
- * @param {number} lineIndex - 0-based line index
- * @param {string} trimmed - Trimmed line content
- */
-function beginFunction(state, name, lineIndex, trimmed) {
-  state.inFunction = true;
-  state.functionName = name;
-  state.functionStart = lineIndex + 1;
-  state.braceDepth = trimmed.includes("{") ? 1 : 0;
-  state.hasReturn = false;
-}
-
-/**
- * Walk shell script lines tracking function boundaries and return statements.
- * @param {string[]} lines - File lines
- * @param {string[]} details - Mutable details array
- * @returns {number} Total violation count
- */
-function walkFunctionsForReturns(lines, details) {
-  let violations = 0;
-  const state = { inFunction: false, functionName: "", functionStart: 0, braceDepth: 0, hasReturn: false };
-
-  for (let i = 0; i < lines.length; i++) {
-    const trimmed = lines[i].trim();
-    const name = matchFunctionDef(trimmed);
-
-    if (name) {
-      violations += checkAndRecordMissingReturn(state, details);
-      beginFunction(state, name, i, trimmed);
-      continue;
-    }
-
-    if (!state.inFunction) continue;
-
-    state.braceDepth += braceDepthDelta(trimmed);
-    if (hasReturnStatement(trimmed)) state.hasReturn = true;
-
-    if (state.braceDepth <= 0) {
-      violations += checkAndRecordMissingReturn(state, details);
-      state.inFunction = false;
-    }
-  }
-
-  violations += checkAndRecordMissingReturn(state, details);
-  return violations;
-}
-
-/**
- * Validate shell script return statements.
- * Checks that functions have explicit return statements (aidevops convention).
- * @param {string} filePath
- * @returns {{ violations: number, details: string[] }}
- */
-function validateReturnStatements(filePath) {
-  const details = [];
-  let violations = 0;
-
-  try {
-    const content = readFileSync(filePath, "utf-8");
-    violations = walkFunctionsForReturns(content.split("\n"), details);
-  } catch {
-    // File read error — skip validation
-  }
-
-  return { violations, details };
-}
-
-/** Patterns for shell constructs where direct $N usage is acceptable. */
-const ALLOWED_POSITIONAL_PATTERNS = [
-  /^\s*shift/,
-  /case\s+.*\$[1-9]/,
-  /getopts/,
-  /"\$@"/,
-  /"\$\*"/,
-];
-
-/**
- * Check if a line uses a shift, case, or getopts pattern that allows direct $N.
- * @param {string} trimmed - Trimmed line content
- * @returns {boolean}
- */
-function isShiftOrCasePattern(trimmed) {
-  return ALLOWED_POSITIONAL_PATTERNS.some((re) => re.test(trimmed));
-}
-
-/** Patterns that indicate currency/pricing/table contexts (false-positives for $N params). */
-const PRICE_TABLE_PATTERNS = [
-  { re: /\$[1-9][0-9.,]/, useStripped: true },            // $28, $1.99, $1,000
-  { re: /\$[1-9]\/(?:mo(?:nth)?|yr|year|day|week|hr|hour)\b/, useStripped: true },  // $5/mo, $9/year
-  { re: /\$[1-9]\s+(?:per|mo(?:nth)?|year|yr|day|week|hr|hour|flat|each|off|fee|plan|tier|user|seat|unit|addon|setup|trial|credit|annual|quarterly|monthly)\b/, useStripped: true },
-  { re: /^\s*\|/, useStripped: false },                    // Markdown table row
-  { re: /\$[1-9]\s*\|/, useStripped: true },               // Pipe-delimited cell
-];
-
-/**
- * Check if a line contains a currency/pricing pattern (false-positive for $N params).
- * @param {string} stripped - Line with escaped dollar signs removed
- * @param {string} rawLine - Original unstripped line
- * @returns {boolean}
- */
-function isPriceOrTablePattern(stripped, rawLine) {
-  return PRICE_TABLE_PATTERNS.some((p) => p.re.test(p.useStripped ? stripped : rawLine));
-}
-
-/**
- * Check whether a trimmed line has a bare positional $N that isn't in a local assignment.
- * @param {string} trimmed - Trimmed line content
- * @returns {boolean}
- */
-function hasBarePositionalParam(trimmed) {
-  return /\$[1-9]/.test(trimmed) && !/local\s+\w+=.*\$[1-9]/.test(trimmed);
-}
-
-/**
- * Check whether stripped content still contains unescaped $N after removing \$N.
- * Also returns false if the remaining $N matches a price/table pattern.
- * @param {string} trimmed - Trimmed line content
- * @param {string} rawLine - Original unstripped line
- * @returns {boolean} true if a real positional param violation exists
- */
-function hasUnescapedPositionalParam(trimmed, rawLine) {
-  const stripped = trimmed.replace(/\\\$[1-9]/g, "");
-  if (!/\$[1-9]/.test(stripped)) return false;
-  if (isPriceOrTablePattern(stripped, rawLine)) return false;
-  return true;
-}
-
-/**
- * Check a single line for positional parameter violations.
- * @param {string} line - Raw line content
- * @param {string} trimmed - Trimmed line content
- * @param {number} lineNum - 1-based line number
- * @param {string[]} details - Mutable details array
- * @returns {number} 0 or 1 (violation count increment)
- */
-function checkPositionalParamLine(line, trimmed, lineNum, details) {
-  if (trimmed.startsWith("#") || !hasBarePositionalParam(trimmed)) return 0;
-  if (isShiftOrCasePattern(trimmed)) return 0;
-  if (!hasUnescapedPositionalParam(trimmed, line)) return 0;
-
-  details.push(`  Line ${lineNum}: direct positional parameter: ${trimmed.substring(0, 80)}`);
-  return 1;
-}
-
-/**
- * Validate positional parameter usage in shell scripts.
- * Checks that $1, $2, etc. are assigned to local variables (aidevops convention).
- * @param {string} filePath
- * @returns {{ violations: number, details: string[] }}
- */
-function validatePositionalParams(filePath) {
-  const details = [];
-  let violations = 0;
-
-  try {
-    const content = readFileSync(filePath, "utf-8");
-    const lines = content.split("\n");
-
-    for (let i = 0; i < lines.length; i++) {
-      violations += checkPositionalParamLine(lines[i], lines[i].trim(), i + 1, details);
-    }
-  } catch {
-    // File read error — skip validation
-  }
-
-  return { violations, details };
-}
+// Shell validator internals extracted to validators.mjs
 
 /**
  * Scan file content for potential secrets.
@@ -1021,78 +534,7 @@ function runShellQualityPipeline(filePath) {
   return { totalViolations, report };
 }
 
-/**
- * Run markdown quality checks on a file.
- * Checks for common issues: trailing whitespace, missing blank lines around
- * code blocks (MD031), consecutive blank lines, broken links.
- * @param {string} filePath
- * @returns {{ totalViolations: number, report: string }}
- */
-/**
- * Check MD031: fenced code blocks should be surrounded by blank lines.
- * @param {string[]} lines - File lines
- * @param {string[]} sections - Mutable report sections array
- * @returns {number} Violation count
- */
-function checkMD031(lines, sections) {
-  let violations = 0;
-  let inCodeBlock = false;
-
-  for (let i = 0; i < lines.length; i++) {
-    if (!/^```/.test(lines[i].trim())) continue;
-
-    if (!inCodeBlock) {
-      if (i > 0 && lines[i - 1].trim() !== "") {
-        sections.push(`  Line ${i + 1}: MD031 — missing blank line before code fence`);
-        violations++;
-      }
-    } else {
-      if (i < lines.length - 1 && lines[i + 1] !== undefined && lines[i + 1].trim() !== "") {
-        sections.push(`  Line ${i + 1}: MD031 — missing blank line after code fence`);
-        violations++;
-      }
-    }
-    inCodeBlock = !inCodeBlock;
-  }
-
-  return violations;
-}
-
-/**
- * Count lines with trailing whitespace.
- * @param {string[]} lines - File lines
- * @param {string[]} sections - Mutable report sections array
- * @returns {number} Violation count
- */
-function checkTrailingWhitespace(lines, sections) {
-  let trailingCount = 0;
-  for (const line of lines) {
-    if (/\s+$/.test(line)) trailingCount++;
-  }
-  if (trailingCount > 0) {
-    sections.push(`  Trailing whitespace on ${trailingCount} line${trailingCount !== 1 ? "s" : ""}`);
-  }
-  return trailingCount;
-}
-
-function runMarkdownQualityPipeline(filePath) {
-  const sections = [];
-  let totalViolations = 0;
-
-  try {
-    const lines = readFileSync(filePath, "utf-8").split("\n");
-    totalViolations += checkMD031(lines, sections);
-    totalViolations += checkTrailingWhitespace(lines, sections);
-  } catch {
-    // File read error — skip
-  }
-
-  const report = sections.length > 0
-    ? `Markdown quality:\n${sections.join("\n")}`
-    : "Markdown checks passed.";
-
-  return { totalViolations, report };
-}
+// Markdown quality checks extracted to quality-pipeline.mjs
 
 /**
  * Check if a tool name is a Write or Edit operation.
@@ -1569,450 +1011,18 @@ function consumeIntent(callID) {
 // pattern to detect violations, and a correction message. Rules can be loaded
 // from a config file or use the built-in defaults.
 
-/**
- * Path to optional user-defined TTSR rules file.
- * JSON array of rule objects: { id, description, pattern, correction, severity }
- * @type {string}
- */
-const TTSR_RULES_PATH = join(AGENTS_DIR, "configs", "ttsr-rules.json");
-
-/**
- * Built-in TTSR rules — enforced by default.
- * Each rule has:
- *   - id: unique identifier
- *   - description: human-readable explanation
- *   - pattern: regex string to detect violations in assistant output
- *   - correction: message injected when violation is detected
- *   - severity: "error" | "warn" | "info"
- *   - systemPrompt: instruction injected into system prompt (preventative)
- *
- * @type {Array<{id: string, description: string, pattern: string, correction: string, severity: string, systemPrompt: string}>}
- */
-const BUILTIN_TTSR_RULES = [
-  {
-    id: "no-glob-for-discovery",
-    description: "Use git ls-files or fd instead of Glob/find for file discovery",
-    pattern: "(?:mcp_glob|Glob tool|use.*\\bGlob\\b.*to find|I'll use Glob)",
-    correction: "Use `git ls-files` or `fd` for file discovery, not Glob. Glob is a last resort when Bash is unavailable.",
-    severity: "warn",
-    systemPrompt: "File discovery: use `git ls-files '<pattern>'` for git-tracked files, `fd` for untracked. NEVER use Glob/find as primary discovery.",
-  },
-  {
-    id: "no-cat-for-reading",
-    description: "Use Read tool instead of cat/head/tail for file reading",
-    pattern: "(?:^|\\s)cat\\s+['\"]?[/~\\w]|\\bhead\\s+-n|\\btail\\s+-n",
-    correction: "Use the Read tool for file reading, not cat/head/tail. These are Bash commands that waste context.",
-    severity: "info",
-    systemPrompt: "Use the Read tool for file reading. Avoid cat/head/tail in Bash — they waste context tokens.",
-  },
-  {
-    id: "read-before-edit",
-    description: "Always Read a file before Edit or Write to existing files",
-    pattern: "(?:I'll edit|Let me edit|I'll write to|Let me write)(?!.*(?:creat|new file|new \\w+ file|generat))(?:(?!I'll read|let me read|I've read|already read).){0,200}$",
-    correction: "ALWAYS Read a file before Edit/Write to an existing file. These tools fail without a prior Read in this conversation. (This rule does not apply when creating new files.)",
-    severity: "error",
-    systemPrompt: "ALWAYS Read a file before Edit or Write to an existing file. These tools FAIL without a prior Read in this conversation. For NEW files, verify the parent directory exists instead.",
-  },
-  {
-    id: "no-credentials-in-output",
-    description: "Never expose credentials, API keys, or secrets in output",
-    pattern: "(?:api[_-]?key|secret|password|token)\\s*[:=]\\s*['\"][A-Za-z0-9+/=_-]{16,}['\"]",
-    correction: "SECURITY: Never expose credentials in output. Use `aidevops secret set NAME` for secure storage.",
-    severity: "error",
-    systemPrompt: "NEVER expose credentials, API keys, or secrets in output or logs.",
-  },
-  {
-    id: "pre-edit-check",
-    description: "Run pre-edit-check.sh before modifying files",
-    pattern: "(?:I'll (?:create|modify|edit|write)|Let me (?:create|modify|edit|write)).*(?:on main|on master)\\b",
-    correction: "Run pre-edit-check.sh before modifying files. NEVER edit on main/master branch.",
-    severity: "error",
-    systemPrompt: "Before ANY file modification: run pre-edit-check.sh. NEVER edit on main/master.",
-  },
-  {
-    id: "shell-explicit-returns",
-    description: "Shell functions must have explicit return statements",
-    pattern: "(?:function\\s+\\w+|\\w+\\s*\\(\\)\\s*\\{)(?:(?!return\\s+[0-9]).){50,}\\}",
-    correction: "Shell functions must have explicit `return 0` or `return 1` statements (SonarCloud S7682).",
-    severity: "warn",
-    systemPrompt: "Shell scripts: every function must have an explicit `return 0` or `return 1`.",
-  },
-  {
-    id: "shell-local-params",
-    description: "Use local var=\"$1\" pattern in shell functions",
-    // Only match bare $N at the start of a line or after whitespace in what looks
-    // like a shell assignment/command context — avoids matching $1 inside prose,
-    // documentation, quoted examples, or tool output from file reads.
-    // Excludes currency/pricing patterns:
-    //   - $[1-9] followed by digits, decimal, or comma (e.g. $28, $1.99, $1,000)
-    //   - $[1-9]/billing-unit (e.g. $5/mo, $9/year) — but NOT $1/config (path)
-    //   - $[1-9] followed by common currency/pricing unit words (per, mo, month, flat, etc.)
-    //   - Escaped dollar signs \$[1-9] (literal dollar in shell strings)
-    pattern: "^\\s+(?:echo|printf|return|if|\\[\\[).*(?<!\\\\)\\$[1-9](?![0-9.,])(?!\\/(?:mo(?:nth)?|yr|year|day|week|hr|hour)\\b)(?!\\s+(?:per|mo(?:nth)?|year|yr|day|week|hr|hour|flat|each|off|fee|plan|tier|user|seat|unit|addon|setup|trial|credit|annual|quarterly|monthly)\\b)(?!.*local\\s+\\w+=)",
-    correction: "Use `local var=\"$1\"` pattern — never use positional parameters directly (SonarCloud S7679).",
-    severity: "warn",
-    systemPrompt: "Shell scripts: use `local var=\"$1\"` — never use $1 directly in function bodies.",
-  },
-];
-
-/**
- * Cached loaded rules (built-in + user-defined).
- * @type {Array<object> | null}
- */
-let _ttsrRules = null;
-
-/**
- * Load TTSR rules: built-in defaults merged with optional user-defined rules.
- * User rules can override built-in rules by matching id.
- * @returns {Array<{id: string, description: string, pattern: string, correction: string, severity: string, systemPrompt: string}>}
- */
-/**
- * Merge user-defined rules into a rules array (mutates the array).
- * Rules with a matching id override the existing entry; new rules are appended.
- * @param {Array<object>} rules - Base rules array (mutated)
- * @param {Array<object>} userRules - User-defined rules to merge
- */
-function mergeUserTtsrRules(rules, userRules) {
-  for (const rule of userRules) {
-    if (!rule.id || !rule.pattern) continue;
-    const existingIdx = rules.findIndex((r) => r.id === rule.id);
-    if (existingIdx >= 0) {
-      rules[existingIdx] = { ...rules[existingIdx], ...rule };
-    } else {
-      rules.push(rule);
-    }
-  }
-}
-
-function loadTtsrRules() {
-  if (_ttsrRules !== null) return _ttsrRules;
-
-  _ttsrRules = [...BUILTIN_TTSR_RULES];
-
-  const userContent = readIfExists(TTSR_RULES_PATH);
-  if (userContent) {
-    try {
-      const parsed = JSON.parse(userContent);
-      if (Array.isArray(parsed)) {
-        mergeUserTtsrRules(_ttsrRules, parsed);
-      }
-    } catch {
-      console.error("[aidevops] Failed to parse TTSR rules file — using built-in rules only");
-    }
-  }
-
-  return _ttsrRules;
-}
-
-/**
- * Check text against a single TTSR rule.
- * @param {string} text - Text to check
- * @param {object} rule - TTSR rule object
- * @returns {{ matched: boolean, matches: string[] }}
- */
-function checkRule(text, rule) {
-  try {
-    const regex = new RegExp(rule.pattern, "gim");
-    const matches = [];
-    let match;
-    while ((match = regex.exec(text)) !== null) {
-      matches.push(match[0].substring(0, 120));
-      if (matches.length >= 3) break; // Cap matches to avoid noise
-    }
-    return { matched: matches.length > 0, matches };
-  } catch {
-    return { matched: false, matches: [] };
-  }
-}
-
-/**
- * Scan text for all TTSR rule violations.
- * @param {string} text - Text to scan
- * @returns {Array<{rule: object, matches: string[]}>}
- */
-function scanForViolations(text) {
-  const rules = loadTtsrRules();
-  const violations = [];
-
-  for (const rule of rules) {
-    const result = checkRule(text, rule);
-    if (result.matched) {
-      violations.push({ rule, matches: result.matches });
-    }
-  }
-
-  return violations;
-}
-
-/**
- * Hook: experimental.chat.system.transform (t1304, t1309)
- *
- * Injects active TTSR rules and intent tracing instruction into the system prompt.
- *
- * Intent tracing (t1309):
- *   Instructs the LLM to include an `agent__intent` field in every tool call,
- *   describing its intent in present participle form. This is the system-prompt
- *   equivalent of oh-my-pi's JSON Schema injection — OpenCode's plugin API does
- *   not expose tool schema modification, so we achieve the same effect via the
- *   system prompt. The field is extracted in tool.execute.before and logged to
- *   the observability DB for debugging and audit trails.
- *
- * TTSR rules (t1304):
- *   Injects active rules as preventative guidance before every LLM call.
- *
- * @param {object} _input - { sessionID?: string, model: Model }
- * @param {object} output - { system: string[] } (mutable)
- */
-async function systemTransformHook(_input, output) {
-  const rules = loadTtsrRules();
-
-  const ruleLines = rules
-    .filter((r) => r.systemPrompt)
-    .map((r) => `- ${r.systemPrompt}`);
-
-  // Intent tracing instruction (t1309) — always injected regardless of TTSR rules
-  const intentInstruction = [
-    "## Intent Tracing (observability)",
-    `When calling any tool, include a field named \`${INTENT_FIELD}\` in the tool arguments.`,
-    "Value: one sentence in present participle form describing your intent (e.g., \"Reading the file to understand the existing schema\").",
-    "No trailing period. This field is used for debugging and audit trails — it is stripped before tool execution.",
-  ].join("\n");
-
-  output.system.push(intentInstruction);
-
-  if (ruleLines.length === 0) return;
-
-  output.system.push(
-    [
-      "## aidevops Quality Rules (enforced)",
-      "The following rules are actively enforced. Violations will be flagged.",
-      ...ruleLines,
-    ].join("\n"),
-  );
-}
-
-/**
- * Extract text content from a Part array.
- * Only extracts text from TextPart objects (type === "text").
- * @param {Array<object>} parts - Array of Part objects
- * @param {object} [options] - Extraction options
- * @param {boolean} [options.excludeToolOutput=false] - Skip tool-result/tool-invocation parts
- * @returns {string} Concatenated text content
- */
-function extractTextFromParts(parts, options = {}) {
-  if (!Array.isArray(parts)) return "";
-  return parts
-    .filter((p) => {
-      if (!p || typeof p.text !== "string") return false;
-      if (p.type !== "text") return false;
-      // When excludeToolOutput is set, skip parts that contain tool output.
-      // Tool results are embedded in assistant messages as text parts whose
-      // content is the serialized tool response. We detect these by checking
-      // for the tool-invocation/tool-result type or by the presence of
-      // toolCallId/toolInvocationId fields that OpenCode attaches.
-      if (options.excludeToolOutput) {
-        if (p.toolCallId || p.toolInvocationId) return false;
-      }
-      return true;
-    })
-    .map((p) => p.text)
-    .join("\n");
-}
-
-/**
- * Cross-turn TTSR dedup state.
- * Tracks which rules have already fired and on which message IDs,
- * preventing the same rule from firing repeatedly on the same content
- * across multiple LLM turns (which caused an infinite correction loop).
- * @type {Map<string, Set<string>>} ruleId -> Set of messageIDs already corrected
- */
-const _ttsrFiredState = new Map();
-
-/**
- * Hook: experimental.chat.messages.transform (t1304)
- *
- * Scans previous assistant outputs for rule violations and injects
- * correction context into the message history. This provides corrective
- * feedback so the model learns from its own violations within the session.
- *
- * Strategy: scan the last N assistant messages (not all — performance).
- * If violations are found, append a synthetic correction message to the
- * message history so the model sees the feedback before generating.
- *
- * Bug fix: Three changes to prevent infinite correction loops:
- * 1. Only scan assistant-authored text, excluding tool output (Read/Bash
- *    results contain code the assistant *read*, not code it *wrote*).
- * 2. Cross-turn dedup — track which rules fired on which messages to
- *    prevent the same rule re-firing on the same content every turn.
- * 3. Skip messages that are themselves synthetic TTSR corrections.
- *
- * @param {object} _input - {}
- * @param {object} output - { messages: { info: Message, parts: Part[] }[] } (mutable)
- */
-/**
- * Filter messages to recent assistant messages, excluding synthetic TTSR corrections.
- * @param {Array} messages - Full message history
- * @param {number} windowSize - Number of recent assistant messages to return
- * @returns {Array} Filtered assistant messages
- */
-function getRecentAssistantMessages(messages, windowSize) {
-  return messages
-    .filter((m) => {
-      if (!m.info || m.info.role !== "assistant") return false;
-      if (m.info.id && m.info.id.startsWith("ttsr-correction-")) return false;
-      return true;
-    })
-    .slice(-windowSize);
-}
-
-/**
- * Collect deduplicated TTSR violations from assistant messages.
- * Applies cross-turn dedup (skips rules already fired on the same message)
- * and within-scan dedup (one violation per rule id per scan).
- * @param {Array} assistantMessages - Messages to scan
- * @returns {Array<{rule: object, matches: string[], msgId: string}>}
- */
-function collectDedupedViolations(assistantMessages) {
-  const allViolations = [];
-
-  for (const msg of assistantMessages) {
-    const msgId = msg.info?.id || "";
-    const text = extractTextFromParts(msg.parts, { excludeToolOutput: true });
-    if (!text) continue;
-
-    const violations = scanForViolations(text);
-    for (const v of violations) {
-      const ruleId = v.rule.id;
-      const firedOn = _ttsrFiredState.get(ruleId);
-      if (firedOn && firedOn.has(msgId)) continue;
-      if (!allViolations.some((av) => av.rule.id === ruleId)) {
-        allViolations.push({ ...v, msgId });
-      }
-    }
-  }
-
-  return allViolations;
-}
-
-/**
- * Record violations in the cross-turn dedup state so they won't re-fire.
- * @param {Array<{rule: object, msgId: string}>} violations
- */
-function recordFiredViolations(violations) {
-  for (const v of violations) {
-    if (!_ttsrFiredState.has(v.rule.id)) {
-      _ttsrFiredState.set(v.rule.id, new Set());
-    }
-    _ttsrFiredState.get(v.rule.id).add(v.msgId);
-  }
-}
-
-/**
- * Build a synthetic correction message for TTSR violations.
- * @param {Array<{rule: object}>} violations
- * @param {string} sessionID
- * @returns {object} Message object to push into output.messages
- */
-function buildCorrectionMessage(violations, sessionID) {
-  const corrections = violations.map((v) => {
-    const severity = v.rule.severity === "error" ? "ERROR" : "WARNING";
-    return `[${severity}] ${v.rule.id}: ${v.rule.correction}`;
-  });
-
-  const correctionText = [
-    "[aidevops TTSR] Rule violations detected in recent output:",
-    ...corrections,
-    "",
-    "Apply these corrections in your next response.",
-  ].join("\n");
-
-  const correctionId = `ttsr-correction-${Date.now()}`;
-
-  return {
-    info: {
-      id: correctionId,
-      sessionID,
-      role: "user",
-      time: { created: Date.now() },
-      parentID: "",
-    },
-    parts: [
-      {
-        id: `${correctionId}-part`,
-        sessionID,
-        messageID: correctionId,
-        type: "text",
-        text: correctionText,
-        synthetic: true,
-      },
-    ],
-  };
-}
-
-async function messagesTransformHook(_input, output) {
-  if (!output.messages || output.messages.length === 0) return;
-
-  const assistantMessages = getRecentAssistantMessages(output.messages, 3);
-  if (assistantMessages.length === 0) return;
-
-  const allViolations = collectDedupedViolations(assistantMessages);
-  if (allViolations.length === 0) return;
-
-  recordFiredViolations(allViolations);
-
-  const sessionID = output.messages[0]?.info?.sessionID || "";
-  output.messages.push(buildCorrectionMessage(allViolations, sessionID));
-
-  qualityLog(
-    "INFO",
-    `TTSR messages.transform: injected ${allViolations.length} correction(s): ${allViolations.map((v) => v.rule.id).join(", ")}`,
-  );
-}
-
-/**
- * Hook: experimental.text.complete (t1304)
- *
- * Detects rule violations post-hoc in completed text parts and flags them.
- * This is observational — it logs violations but does not modify the output
- * (the text has already been shown to the user). The log data feeds into
- * quality metrics and pattern tracking.
- *
- * @param {object} input - { sessionID: string, messageID: string, partID: string }
- * @param {object} output - { text: string } (mutable)
- */
-async function textCompleteHook(input, output) {
-  if (!output.text) return;
-
-  const violations = scanForViolations(output.text);
-  if (violations.length === 0) return;
-
-  // Log violations for observability
-  for (const v of violations) {
-    qualityLog(
-      v.rule.severity === "error" ? "ERROR" : "WARN",
-      `TTSR violation [${v.rule.id}]: ${v.rule.description} (session: ${input.sessionID}, message: ${input.messageID})`,
-    );
-  }
-
-  // Append violation markers as comments at the end of the text
-  // so the model can see them in subsequent turns
-  const markers = violations.map((v) => {
-    const severity = v.rule.severity === "error" ? "ERROR" : "WARN";
-    return `<!-- TTSR:${severity}:${v.rule.id} — ${v.rule.correction} -->`;
-  });
-
-  output.text = output.text + "\n" + markers.join("\n");
-
-  // Record pattern for tracking
-  const patternTracker = join(SCRIPTS_DIR, "pattern-tracker-helper.sh");
-  if (existsSync(patternTracker)) {
-    const ruleIds = violations.map((v) => v.rule.id).join(",");
-    run(
-      `bash "${patternTracker}" record "TTSR_VIOLATION" "rules: ${ruleIds}" --tag "ttsr" 2>/dev/null`,
-      5000,
-    );
-  }
-}
+const {
+  systemTransformHook,
+  messagesTransformHook,
+  textCompleteHook,
+} = createTtsrHooks({
+  agentsDir: AGENTS_DIR,
+  scriptsDir: SCRIPTS_DIR,
+  readIfExists,
+  qualityLog,
+  run,
+  intentField: INTENT_FIELD,
+});
 
 // ---------------------------------------------------------------------------
 // Main Plugin Export
